# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: InsurRouter API - Two-wheeler insurance application routing engine
servers:
  - url: /api
    description: Base API path
...
security:
  - serviceKey: []
paths:
...
  /dms/showrooms/{showroomId}/sync:
    post:
      operationId: syncShowroomDms
      tags: [dms]
      summary: Pull a showroom's deals from its DMS into the local mirror
      description: >
        Read-only. Walks every active DMS account on the showroom and upserts
        each deal. Only fetches a full record when the summary shows something
        moved, so a routine sync on a quiet day is one list call.
      parameters:
        - name: showroomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: One result per DMS account synced
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/DmsSyncResult"
                required: [results]
        "404":
          description: No such showroom, or it has no active DMS account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: DMS unreachable or not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dms/showrooms/{showroomId}/panel:
    get:
      operationId: getShowroomPanel
      tags: [dms]
      summary: The insurers this showroom can actually place business with
      description: >
        Joins the global `providers` list (how we reach an insurer) to the
        dealer's own arrangement with them (quota, payout, turnaround). Quota
        and eligibility are computed by the shared quoting module, so the rules
        are identical to the ones the dealer portal applies.
      parameters:
        - name: showroomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Panel entries, by dealer code then insurer
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: "#/components/schemas/PanelEntry"
                required: [entries]
        "404":
          description: No such showroom
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /dms/worklist:
    get:
      operationId: getShowroomWorklist
      tags: [dms]
      summary: Every mirrored deal with its DMS-versus-DDMS reconciliation
      description: >
        The owner's console. Each row carries the dealer system's view and ours
        side by side, plus how they differ and the single action that would
        close the gap. Reconciliation is computed on read, never stored, so it
        cannot go stale.

        Reads only the mirror and never calls the DMS, so the console still
        renders when the dealer's ERP is busy or down.

        `showroomId` is a query parameter rather than a path segment on purpose.
        Orval emits `<Operation>Params` for path parameters *and* for query
        parameters, so an operation carrying both produces the same exported
        name twice and the api-zod barrel fails to compile. `listApplications`
        already follows this query-only shape for the same reason — do not
        "tidy" it into /dms/showrooms/{id}/worklist without checking codegen.
      parameters:
        - name: showroomId
          in: query
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: false
          description: Filter on the DMS status, e.g. AWAITING_INSURANCE
          schema:
            type: string
        - name: includeDisappeared
          in: query
          required: false
          description: Include deals the DMS has stopped listing. Off by default.
          schema:
            type: boolean
      responses:
        "200":
          description: Worklist rows plus a summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    $ref: "#/components/schemas/WorklistSummary"
                  rows:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorklistRow"
                required: [summary, rows]
        "404":
          description: No such showroom
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Ingest ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    serviceKey:
      type: apiKey
      in: header
      name: x-api-key
      description: >
        Shared secret proving the caller is one of our own processes. The
        api-server also accepts it as `Authorization: Bearer <key>`; both are
        the same credential.

        The two web frontends never hold it — each Vite dev server injects the
        header into its `/api` proxy, so it stays in Node and out of the bundle.

        It authenticates a process, not a person. Which showroom is asking still
        arrives in the request rather than being derived from a session, so this
        does not yet enforce the owner scoping the schema is shaped for.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    PanelEntry:
      type: object
      description: >
        One insurer on one outlet's panel. `providers` says how to reach an
        insurer; this says what this dealer's arrangement with them is. Neither
        is derivable from the other.
      properties:
        insurerCode: { type: string }
        insurerName: { type: string }
        shortName: { type: string }
        providerId:
          type: integer
          description: The providers row, so a caller can execute against it.
        dealerCode:
          type: string
          description: >
            Which DMS account's arrangement this is. A showroom holding two
            brands has two panels and both appear — a deal from one brand
            cannot be placed on the other brand's panel.
        route:
          type: string
          enum: [BROKER, DIRECT_AGENT]
          description: >
            Taken from the DMS account's insuranceChannel, never stored on the
            panel row, so the two cannot disagree.
        integration:
          type: string
          enum: [API, PORTAL]
          description: >
            What this dealer can reach. Not the same as the provider's own
            execution mode: a broker platform can front an insurer through a
            web portal even where that insurer publishes an API.
        odBaseRate: { type: number }
        quotaPolicies: { type: integer }
        quotaConsumed: { type: integer }
        payoutRate: { type: number }
        slaMinutes: { type: integer }
        isEnabled: { type: boolean }
        quota:
          type: object
          properties:
            remaining: { type: integer }
            utilisation: { type: number }
            exhausted: { type: boolean }
            tight:
              type: boolean
              description: Near the limit — worth warning about, still selectable.
          required: [remaining, utilisation, exhausted, tight]
        eligible: { type: boolean }
        reasons:
          type: array
          description: >
            Why an insurer is excluded, as a list rather than a boolean. A
            dealer who cannot see the reason will not trust the recommendation,
            and an unexplained recommendation is worse than none.
          items: { type: string }
      required:
        [insurerCode, insurerName, shortName, providerId, dealerCode, route,
         integration, odBaseRate, quotaPolicies, quotaConsumed, payoutRate,
         slaMinutes, isEnabled, quota, eligible, reasons]
    DmsSyncResult:
      type: object
      description: What one sync of one DMS account did.
      properties:
        showroomId: { type: integer }
        dealerCode: { type: string }
        seen:
          type: integer
          description: Deals the DMS listed.
        added: { type: integer }
        changed: { type: integer }
        unchanged: { type: integer }
        disappeared:
          type: integer
          description: >
            Mirrored deals the DMS no longer lists. Marked, never deleted — a
            deal vanishing is itself information.
        startedAt: { type: string }
        finishedAt: { type: string }
        durationMs: { type: integer }
      required:
        [showroomId, dealerCode, seen, added, changed, unchanged, disappeared,
         startedAt, finishedAt, durationMs]
    ReconcileState:
      type: string
      enum: [IN_SYNC, AHEAD, BEHIND, CONFLICT, NOT_STARTED]
      description: >
        How our record compares to the dealer's system for one deal.
        AHEAD means we did the work and their system does not know yet — the
        normal state under a read-only integration, and a task rather than an
        error. BEHIND means somebody worked outside DDMS. CONFLICT means two
        different policy numbers on one vehicle.
    WorklistRow:
      type: object
      properties:
        dealId: { type: string }
        dealerCode: { type: string }
        showroomId: { type: integer }
        showroomCode: { type: ["string", "null"] }
        customerName: { type: ["string", "null"] }
        modelDescription: { type: ["string", "null"] }
        chassisNo: { type: ["string", "null"] }
        bookingDate: { type: ["string", "null"] }
        dms:
          type: object
          description: What the dealer's own system currently believes.
          properties:
            status: { type: string }
            policyNo: { type: ["string", "null"] }
            insurerCode: { type: ["string", "null"] }
            regNo: { type: ["string", "null"] }
          required: [status]
        ddms:
          type: object
          description: What we know, which may be more recent.
          properties:
            applicationId: { type: ["integer", "null"] }
            applicationStatus: { type: ["string", "null"] }
            policyNumber: { type: ["string", "null"] }
            policySimulated:
              type: boolean
              description: True when the number came from a stub, not an insurer.
            providerName: { type: ["string", "null"] }
          required: [policySimulated]
        reconcile:
          $ref: "#/components/schemas/ReconcileState"
        reconcileNote:
          type: ["string", "null"]
          description: Plain-English statement of the difference.
        actionRequired:
          type: ["string", "null"]
          description: The one thing to do next, when there is one.
        daysInStatus:
          type: integer
          description: >
            Days since the DMS status last moved. This is the number nobody has
            time to work out by hand, and the reason a mirror exists at all.
        lastSyncedAt: { type: string }
        disappearedFromDms: { type: boolean }
      required:
        [dealId, dealerCode, showroomId, dms, ddms, reconcile, daysInStatus,
         lastSyncedAt, disappearedFromDms]
    WorklistSummary:
      type: object
      properties:
        total: { type: integer }
        byReconcile:
          type: object
          additionalProperties: { type: integer }
        byDmsStatus:
          type: object
          additionalProperties: { type: integer }
        needsAction:
          type: integer
          description: Rows with an action attached. What the console leads with.
        oldestDaysInStatus: { type: integer }
        lastSyncedAt: { type: ["string", "null"] }
      required: [total, byReconcile, byDmsStatus, needsAction, oldestDaysInStatus]
...