# lib/api-zod/src/generated/types/panelEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * InsurRouter API - Two-wheeler insurance application routing engine
 * OpenAPI spec version: 0.1.0
 */
import type { PanelEntryIntegration } from './panelEntryIntegration';
import type { PanelEntryQuota } from './panelEntryQuota';
import type { PanelEntryRoute } from './panelEntryRoute';

/**
 * One insurer on one outlet's panel. `providers` says how to reach an insurer; this says what this dealer's arrangement with them is. Neither is derivable from the other.
 */
export interface PanelEntry {
  insurerCode: string;
  insurerName: string;
  shortName: string;
  /** The providers row, so a caller can execute against it. */
  providerId: number;
  /** Which DMS account's arrangement this is. A showroom holding two brands has two panels and both appear — a deal from one brand cannot be placed on the other brand's panel. */
  dealerCode: string;
  /** Taken from the DMS account's insuranceChannel, never stored on the panel row, so the two cannot disagree. */
  route: PanelEntryRoute;
  /** What this dealer can reach. Not the same as the provider's own execution mode: a broker platform can front an insurer through a web portal even where that insurer publishes an API. */
  integration: PanelEntryIntegration;
  odBaseRate: number;
  quotaPolicies: number;
  quotaConsumed: number;
  payoutRate: number;
  slaMinutes: number;
  isEnabled: boolean;
  quota: PanelEntryQuota;
  eligible: boolean;
  /** Why an insurer is excluded, as a list rather than a boolean. A dealer who cannot see the reason will not trust the recommendation, and an unexplained recommendation is worse than none. */
  reasons: string[];
}